feat: 3차 디자인 수정
